refactor: backend refactored to provide butter structure
