+++ backend/prisma/schema.prisma
@@ -14,7 +14,6 @@ datasource db {
 }
 
 enum Role {
-  ADMIN
   TEACHER
   STUDENT
   PRINCIPAL
