--- backend/src/admin/admin.module.ts
@@ -1,22 +1,0 @@
-import { Module } from '@nestjs/common';
-import { AdminController } from './admin.controller';
-import { AdminService } from './admin.service';
-import { PrismaModule } from '../prisma/prisma.module';
-import { ClassModule } from '../class/class.module';
-import { AcademyModule } from '../academy/academy.module';
-import { JwtModule } from '@nestjs/jwt';
-
-@Module({
-  imports: [
-    PrismaModule,
-    ClassModule,
-    AcademyModule,
-    JwtModule.register({
-      secret: process.env.JWT_SECRET,
-      signOptions: { expiresIn: '1d' },
-    }),
-  ],
-  controllers: [AdminController],
-  providers: [AdminService],
-})
-export class AdminModule {}
