--- backend/src/admin/dto/index.ts
@@ -1,6 +1,0 @@
-export * from './create-student.dto';
-export * from './create-teacher.dto';
-export * from './create-class.dto';
-export * from './reset-password.dto';
-export * from './withdrawal-stats.dto';
-export * from './create-academy.dto';
