--- backend/src/admin/dto/reset-password.dto.ts
@@ -1,8 +1,0 @@
-import { IsString } from 'class-validator';
-import { ApiProperty } from '@nestjs/swagger';
-
-export class ResetPasswordDto {
-  @ApiProperty({ example: 'new_secure_password', description: '새 비밀번호' })
-  @IsString()
-  newPassword: string;
-}
