--- backend/src/admin/entities.ts
@@ -1,34 +1,0 @@
-import { ApiProperty } from '@nestjs/swagger';
-
-export class TeacherEntity {
-  @ApiProperty()
-  id: number;
-
-  @ApiProperty()
-  name: string;
-
-  @ApiProperty()
-  email: string;
-}
-
-export class StudentEntity {
-  @ApiProperty()
-  id: number;
-
-  @ApiProperty()
-  name: string;
-
-  @ApiProperty()
-  email: string;
-}
-
-export class ClassEntity {
-  @ApiProperty()
-  id: number;
-
-  @ApiProperty()
-  name: string;
-
-  @ApiProperty()
-  teacherId: number;
-}
