+++ backend/src/auth/types/auth.types.ts
@@ -1,4 +1,4 @@
-export type UserRole = 'student' | 'teacher' | 'admin';
+export type UserRole = 'student' | 'teacher' | 'principal';
 
 export interface JwtPayload {
   sub: string;
