+++ backend/src/types/class.types.ts
@@ -1,16 +1,28 @@
 export interface CreateClassDto {
-  name: string;
+  className: string;
   description: string;
+  maxStudents: number;
+  tuitionFee: number;
+  teacherId: number;
+  academyId: number;
   dayOfWeek: DayOfWeek;
+  level: string;
   startTime: string;
   endTime: string;
-  maxStudents: number;
-  teacherId: number;
+  startDate: string; // UTC ISO 문자열
+  endDate: string; // UTC ISO 문자열
 }
 
 export interface UpdateClassDto extends Partial<CreateClassDto> {}
 
-export type DayOfWeek = 'MONDAY' | 'TUESDAY' | 'WEDNESDAY' | 'THURSDAY' | 'FRIDAY' | 'SATURDAY' | 'SUNDAY';
+export type DayOfWeek =
+  | 'MONDAY'
+  | 'TUESDAY'
+  | 'WEDNESDAY'
+  | 'THURSDAY'
+  | 'FRIDAY'
+  | 'SATURDAY'
+  | 'SUNDAY';
 
 export interface EnrollmentDto {
   studentId: number;
